+++ core/src/test/java/arez/CountingProcedure.java
@@ -1,0 +1,19 @@
+package arez;
+
+class CountingProcedure
+  implements Procedure
+{
+  private int _callCount;
+
+  @Override
+  public void call()
+    throws Throwable
+  {
+    _callCount++;
+  }
+
+  int getCallCount()
+  {
+    return _callCount;
+  }
+}